+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Add support for Compositing and Blending Level 1
   - <https://github.com/vivliostyle/vivliostyle.js/issues/148>
   - Spec: [Compositing and Blending Level 1](https://drafts.fxtf.org/compositing-1/)
+- pseudo element or page margin box with `content:url()` behave as a replaced element
+  - Spec: [Inserting and replacing content with the content property](https://drafts.csswg.org/css-content/#content-property)
 
 ### Fixed